feat(config): Add remote build cache provider types
