+++ packages/@expo/cli/src/run/ios/XcodeBuild.types.ts
@@ -1,5 +1,6 @@
+import { RemoteBuildCacheProvider } from '@expo/config';
+
 import { OSType } from '../../start/platforms/ios/simctl';
-import { RemoteBuildCacheProvider } from '../../utils/remote-build-cache-providers/types';
 import { BundlerProps } from '../resolveBundlerProps';
 
 export type XcodeConfiguration = 'Debug' | 'Release';
