+++ packages/@expo/config/src/index.ts
@@ -4,3 +4,4 @@ export * from './Config';
 export * from './Config.types';
 export * from './getExpoSDKVersion';
 export * from './Errors';
+export * from './remoteBuildCache';
